LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-outside-ellipse-000-expected.txt


FAIL No arguments - inline assert_equals: expected "ellipse()" but got "ellipse(at 50% 50%)"
FAIL Radial arguments - inline assert_equals: expected "ellipse(50px 50px)" but got "ellipse(50px 50px at 50% 50%)"
PASS Position argument - inline
PASS Radial and position argument - inline
FAIL No arguments - computed assert_equals: expected "ellipse()" but got "ellipse(at 50% 50%)"
FAIL Radial arguments - computed assert_equals: expected "ellipse(50px 50px)" but got "ellipse(50px 50px at 50% 50%)"
PASS Position argument - computed
PASS Radial and position argument - computed
